% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_em_func.r
\docType{data}
\name{example}
\alias{example}
\title{A data example}
\value{
A list containing
\item{X}{bulk gene expression (gene x subject x measure).}
\item{W}{subject-specific cell type fraction (subject x measure x cell type).}
}
\description{
A data list for demonstration.
}
\examples{

data(example)

}
