% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_em_func.r
\name{get_network}
\alias{get_network}
\title{Making a network plot for co-expressed genes}
\usage{
get_network(alpha, W, cell_type, cor_cutoff = 0.9, gene_color = "#CD2836")
}
\arguments{
\item{alpha}{the deconvolved cell-type-specific gene expression (gene x cell type x individual).}

\item{W}{subject-specific cell type fraction (individual x measure x cell type).}

\item{cell_type}{a numeric index indicating which cell type to be plotted.}

\item{cor_cutoff}{the threshold of weighted correlation for linkages to be plotted.}

\item{gene_color}{the color to depict genes in the network. When a single value is provided, all genes are plotted in the same color.
If a vector of length of genes is provided, each gene will be plotted with the provided color.}
}
\value{
An interactive circle network plot made by edgebundleR to visualize the weighted co-expression of genes based on deconvolved expression.
}
\description{
It makes a circle network plot for co-expressed genes based on the weighted correlation of the deconvolved cell-type-specific expression. The correlation is
weighted by the average cell type fraction per person.
}
